python: fix attribute deletion segfault
Previous commit 93d2ef8b730bc2659e348bce9f7ecfc902600b9d introduced a
bug where an attempt to delete a computed attribute results in a
segmentation fault.

The setters in C bindings have been modified to check whether the new
parameter is NULL and throw an exception if it is.

Signed-off-by: Martin Milata <mmilata@redhat.com>
Signed-off-by: Richard Marko <rmarko@redhat.com>