Add type-agnostic API
Type-agnostic API means that we expose several functions that operate on
frames (or threads or stacktraces) regardless of their problem type.
This is accomplished by looking at the type member of each structure
(introduced in previous patch) and using it to look up the right
function for the type.

Furthermore, the implementations of the type-agnostic functions can
themselves work on multiple structure types, which can prevent some code
duplication.

A sr_stacktrace_to_short_text function is provided as an example,
closing #64. Related to #63 and related to #66 as the requested
functionality should be easier to implement now.

Note:

I've fixed an obvious typo in lib/utils.h
- (assert((type > 0) && (type) < SR_REPORT_NUM
+ (assert((type > SR_REPORT_INVALID) && (type) < SR_REPORT_NUM

Jakub Filak

Signed-off-by: Martin Milata <mmilata@redhat.com>
Signed-off-by: Jakub Filak <jfilak@redhat.com>