python: treat UINT64_MAX as None
Note: this patch depends on "[SATYR PATCH 7/7] tests for json
deserialization of stacktraces" as it modifies one of the tests.

-->8-->8-->8-->8-->8-->8-->8-->8-->8-->8-->8-->8-->8-->8-->8-->8-->8--
Some structures (namely struct sr_core_frame) use (uint64_t)-1 to denote
unknown value. This is difficult to compare with in python, so this
patch introduces translation of this value to None in uint64 accessor
functions.

Signed-off-by: Martin Milata <mmilata@redhat.com>
Signed-off-by: Richard Marko <rmarko@redhat.com>