python: factor out stacktrace object creation
The code is repeated at least two times and we'll need it once again.
It is still essentially the same for all four stacktrace types and could
probably be rewritten in a generic way but I'd rather not do it now for
the sake of size of this patchset.

Related to #86.

Signed-off-by: Martin Milata <mmilata@redhat.com>
Signed-off-by: Richard Marko <rmarko@redhat.com>