koops: recognize multiple x86_64 stacks
Closes #73. Closes #108.

I decided to implement this as an attribute of koops frames as opposed
to having each stack as a separate thread for two reasons:

1) uReport backward compatibility - when a report comes with the
"frames" element only (i.e. no special stack), does it mean that the
stacks are concatenated together (old version) or that there's really no
special stack (new version)?

2) Clustering, which we currently do on a single crash thread. Based on
my understanding and a conversation with a kernel guy, there's no simple
rule based on which one of the stacks can be chosen as relevant.

The first issue can be addressed in next version of the uReport format.
The second issue should be addressed based on concrete suggestions from
kernel developers.

Signed-off-by: Martin Milata <mmilata@redhat.com>