Demangle C++ symbols, closes #130
C++ symbols are now demangled during stack unwinding. A function is
provided to the library user to do this.

Please note that the demangling algorithm is different from the one used
in c++filt (which uses libiberty which is not available as a shared
library). E.g. for "_ZN9wikipedia7article8print_toERSo":
c++filt: wikipedia::article::print_to(std::basic_ostream<char, std::char_traits<char> >&)
satyr:   wikipedia::article::print_to(std::ostream&)

Also note that we have no way to find out if the symbol is actually
mangled, so the demangling function is attempted on everthing that
starts with "_Z".

Signed-off-by: Martin Milata <mmilata@redhat.com>