Implement duphash for kernel oopses
That is (partially) compatible with the algorithm in ABRT <= 2.1.10.

Note that the original hashing procedure had one special case that is
not implemented here: for stack traces where the string "Call Trace:" is
not present but the string "WARNING: at" is, the next two tokens after
"WARNING: at" were used to compute the hash:

/* Special-case: if the first line is of form:
 * WARNING: at net/wireless/core.c:614 wdev_cleanup_work+0xe9/0x120 [cfg80211]() (Not tainted)
 * then hash only "file:line func+ofs/len" part.
 */

Related to rhbz#1040900.

Signed-off-by: Martin Milata <mmilata@redhat.com>