tests: change the way satyr module is imported
In order for satyr objects to be picklable, their types/constructors
have to be available in the "satyr" module. However, the current module
was named _satyr which resulted in following error when trying to pickle
something:

PicklingError: Can't pickle <type 'satyr.DistancesPart'>: it's not found as satyr.DistancesPart

Related to #143.

Signed-off-by: Martin Milata <mmilata@redhat.com>