Disable fingerprinting on non-x86_64
The fingerprinting was disabled on non-x86_64 arches by means of
disabling the disassembly routines which failed and therefore the whole
fingerprinting failed. However, x86_64-specific PLT table parsing was
done before using the disassembler, which was observed to cause a
segfault on ppc64. This disables the fingerprinting at the
fingerprinting procedure level.

Related to rhbz#1052402.

Signed-off-by: Martin Milata <mmilata@redhat.com>