Support custom uReport elements
We need this for 'machine id' element which should be present only if
users wishes to it.

The custom entries are string based, key value pairs (JSON string
elements).

The custom entries are enclosed in an envelope object called 'custom' in
order to be able to distinguish between the custom and the standard elements.

Example:

"custom" : {   "machineid" : "deadbeaf"
           }

Fixes #171

Signed-off-by: Jakub Filak <jfilak@redhat.com>