abrt: refactorize unwinding from core hook
ABRT needs to save files from its core hook with O_EXCL and with right
mode and owner and there might be more constraints ABRT needs to
respect in future.

This patch breaks granularity of the abrt functions to allow ABRT to save the
core backtrace data in the way it need by providing a function returning
the data instead of directly saving the data to a file.

Signed-off-by: Jakub Filak <jfilak@redhat.com>