gdb: parse and store Thread ID (LWP value)
We only need that value to identify crash thread. There is no other
reason to parse it I can come up with.

Do not use it in 'sr_gdb_thread_append_to_str()' (not even in verbose
mode) because it would be included in the "truncated backtrace"
('sr_gdb_stacktrace_to_short_text()').

Do not use in 'sr_gdb_thread_cmp()' as that value has local scope and is
different for each program run.

Related: #227

Signed-off-by: Jakub Filak <jfilak@redhat.com>