gdb: use TID to identify Crash thread in GDB backtrace
It is the one of the possible approaches how to identify the crash
thread and it seems to be more accurate than searching for the crash
frame as there might be more threads having the crash frame on the top
of their stacks.

The new stacktrace example has a wrong function name in Frame #0 in
Thread #1, otherwise, the frame is exactly the crash frame in the
backtrace header (bugzilla.redhat.com/1233787).

Related: #227

Signed-off-by: Jakub Filak <jfilak@redhat.com>