Introduce 'serial' field in uReport
The very first idea was send just debut information, if the crash occured first
time or not. After, we decided to send whole information about crash occurrences
(count element from dump dir) as serial. The reason is the crash occurrences
give us some security check or call it uReports validation. All uReports should
have growing serial number. Hit rate of the uReport serial number should be
decreasing with increasing value of the serial.

It is better to send more detailed information (serial) then boolean value
(debut). It may be useful in the future. We are not aware of any security issue
linked to sending the local occurrence counter's value. The idea of the 'debut'
bit was introduced because we were afraid of potential misinterpretation of the
number - e.g. one could think the number should be added to server crash
counter.

Related abrt/faf#484

Signed-off-by: Jakub Filak <jfilak@redhat.com>