elfutils: run hook unwinding under unprivileged user
We need to divide the unwinding into too steps because we want to run
elfutils under unprivileged users in abrt-hook-ccpp.

This commit splits the unwinding into prepare and generate parts. The
prepare part includes attaching the process using PTRACE - this part
must be run under privileged users. The generate part includes unwinding.

With some modifications to elfutils, we might be able to run all
elfutils function under an unprivileged user.

Signed-off-by: Adam Sulc <xmineral@seznam.cz>
Signed-off-by: Jakub Filak <jfilak@redhat.com>