Replace bundled JSON parser with json-c
This commit drops the bundled JSON parser from
https://github.com/udp/json-parser/ and replaces it with json-c, an
external library. It’s already used in libreport, so we’re not breaking
new ground here by adding alien technology.

With this, we don’t have to care about testing the actual parser,
dealing with any security issues that may arise or otherwise maintaining
the code. All that and less code to compile.

Macros were removed where it didn’t add too much code, because otherwise
debugging is a massive pain in the hindside. The craziest ones remain.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>