tests: Don't _FORTIFY_SOURCE for dump_core.c
The stand-alone binary that is supposed to dump core to be used in tests
needs to be built without optimization to provide a sufficient number of
frames in the core dump. The disabled optimization clashes with
-D_FORTIFY_SOURCE=2 that rpmbuild sets when running make rpm, causing the
testsuite build to fail with a misleading error (__open_too_many_args ()).
This commit turns off _FORTIFY_SOURCE for the core-dumping binary.

Resolves:
https://github.com/abrt/satyr/issues/311

Signed-off-by: Michal Fabik <mfabik@redhat.com>