spec: Output logs from failed tests
After the changes started in c465acea, spec-driven builds no longer
output the logs for failed tests as the logs are now stored in .log
files named after the corresponding test C source (and binary), rather
than in tests/testsuite.dir. This, however, doesn't apply to Python
bindings tests, which is why the `find` on tests/testsuite.dir is
retained.

Signed-off-by: Michal Fabik <mfabik@redhat.com>