Add support for fine-grained error location lines in Python tracebacks
See: PEP 657

Python 3.11 makes it easier to identify where on the line a problem
occurred. This is achieved with the error location lines.
We need to skip such lines when parsing the traceback, otherwise
we won't identify the exception name correctly (See: rhbz#2137473).

Example:

Traceback (most recent call last):
  File "test.py", line 2, in <module>
    x['a']['b']['c']['d'] = 1
    ~~~~~~~~~~~^^^^^
TypeError: 'NoneType' object is not subscriptable

Signed-off-by: Michal Srb <michal@redhat.com>