Find crash thread before stacktrace is normalized
When we normalize a stacktrace, we sometimes remove certain
frames. The problem is when crash frame is one of those
removed frames. In such cases, we simply cannot (reliably)
find the crash thread.

Therefore, before we normalize the stacktrace, we look
for the crash thread and remember its id. We can then
use that id later, if needed.

Related: rhbz#2168223
Fixes: #337

Signed-off-by: Michal Srb <michal@redhat.com>